--- rendererWorkStation/schematics.py
@@ -1,8 +1,0 @@
-import math
-
-import sys
-sys.path.append('/3D-renderer/rendererWorkStation')
-
-#from helperFunctions import *
-from rendererWorkStation.ratioSchematic import *
-from cubeSchematic import *
